+++ src/EventListener/SmsControllerEventListener.php
@@ -5,28 +5,31 @@ declare(strict_types=1);
 namespace BikeShare\EventListener;
 
 use BikeShare\Db\DbInterface;
-use BikeShare\Event\SmsDuplicateDetectedEvent;
+use BikeShare\Notifier\AdminNotifier;
 use BikeShare\SmsConnector\SmsConnectorInterface;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpKernel\Event\ControllerEvent;
-use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 class SmsControllerEventListener
 {
     private DbInterface $db;
     private SmsConnectorInterface $smsConnector;
-    private EventDispatcherInterface $eventDispatcher;
+    private AdminNotifier $adminNotifier;
+    private TranslatorInterface $translator;
     private LoggerInterface $logger;
 
     public function __construct(
         DbInterface $db,
         SmsConnectorInterface $smsConnector,
-        EventDispatcherInterface $eventDispatcher,
+        AdminNotifier $adminNotifier,
+        TranslatorInterface $translator,
         LoggerInterface $logger
     ) {
         $this->db = $db;
         $this->smsConnector = $smsConnector;
-        $this->eventDispatcher = $eventDispatcher;
+        $this->adminNotifier = $adminNotifier;
+        $this->translator = $translator;
         $this->logger = $logger;
     }
 
@@ -56,9 +59,9 @@ class SmsControllerEventListener
         if ($result->rowCount() >= 1) {
             // sms already exists in DB, possible problem
             $this->logger->error("SMS already exists in DB", compact('sms_uuid'));
-            $this->eventDispatcher->dispatch(
-                new SmsDuplicateDetectedEvent($sms_uuid),
-                SmsDuplicateDetectedEvent::NAME
+            $this->adminNotifier->notify(
+                $this->translator->trans('Problem with SMS') . $sms_uuid,
+                false
             );
         } else {
             $this->db->query(
